预览传出的pts由32位转成64位
git-svn-id: svn://192.168.2.188/qtcms@684 894f9759-fa2a-45cf-9159-0a04a166245e

+++ src/functionlibs/BubbleProtocol/StreamProcess.cpp
@@ -316,7 +316,7 @@ void StreamProcess::analyzePreviewStream()
 				GetWidthHeight(liveStreamInfo->pData, qToBigEndian(liveStreamInfo->uiLength), &width, &height);	
 
 				mStreamInfo.insert("channel", liveStreamInfo->cChannel);
-				mStreamInfo.insert("pts", qToBigEndian(pBubble->uiTicket)*1000);
+				mStreamInfo.insert("pts", (unsigned long long)qToBigEndian(pBubble->uiTicket)*1000);
 				mStreamInfo.insert("length", qToBigEndian(liveStreamInfo->uiLength));
 				mStreamInfo.insert("data", (int)(liveStreamInfo->pData));
 				mStreamInfo.insert("frametype", liveStreamInfo->cType);
